Move SerializePositionalToKwargsBridge to the plugin itself
