+++ lib/manageiq/schema/serialize_positional_to_kwargs_bridge.rb
@@ -1,0 +1,19 @@
+module ManageIQ
+  module Schema
+    module SerializePositionalToKwargsBridge
+      def serialize(*args, **options)
+        return super if Rails.version >= "7.1"
+
+        # For rails 7.0.x, convert 7.1+ kwargs for coder/type into the positional argument
+        # class_name_or_coder
+        if options[:coder]
+          args << options.delete(:coder)
+        elsif options[:type]
+          args << options.delete(:type)
+        end
+
+        super(*args, **options)
+      end
+    end
+  end
+end
